Move 'get_default_date' to the module where it's used, to improve code cohesion
